Implement half of the game loop. Move functions around.
button action functions moved from aux.py to game class. This is so
game does not have to be passed as a parameter.

Add to game class attributes gameStates, stateName, and state. These are
used to manage the current game state. e.g. mainMenu, start, guessing, etc...
The game loop is halfway implemented (as in, the game starts, the menus do
the appropriate things, the boards are drawn, and ships can be placed.

ShipPlacement() should be rewritten to conform with the general flow of
pygame.