Fix aspect ratio problem.
When starting the game on a monitor with a nontraditional aspect ratio,
game boards do not scale correctly to fix onto the screen.
Current solution fixes the aspect ratio by grabbing the height and scaling
the width off that. Further solutions should probably involve grabbing
the min of height and width / CONSTANT to adapt for taller aspect ratios
like 4:3